+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/Orange-OpenSource/ouds-ios-design-system-toolbox/compare/1.1.0...develop)
 
+### Added
+
+- [Library] `password input` component (Orange-OpenSource/ouds-ios#997)
+
 ### Changed
 
 - Rename of *Orange Business Tools*  theme to *Orange Compact* (Orange-OpenSource/ouds-ios#1292)
@@ -18,7 +22,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- [Library] `password input` component (Orange-OpenSource/ouds-ios#997)
 - [Library] Add flag to let `button` takes the screen full width (Orange-OpenSource/ouds-ios#1252)
 - [Library] Stop animation of switch if defined in accessibility settings (Orange-OpenSource/ouds-ios#660)
 - [Library] Display in logs components versions